Refactor Logo logoc Controller and Model

+++ backend/controller/v1/Logo.controller.js
@@ -1,4 +1,5 @@
-import Logo from "../../models/Logo.model"
+import { deleteImage } from "../../middleware/Cloudinary-api.js"
+import Logo from "../../models/Logo.model.js"
 
 export const GetAllLogo = async(req,res)=>{
     try {
@@ -12,12 +13,12 @@ export const GetAllLogo = async(req,res)=>{
 
 export const AddLogo = async(req,res) => {
     try {
-        const {imageUrl} = req.body
-        if(!imageUrl) {
+        const {image} = req.body
+        if(!image) {
             return res.status(400).json({message:"Images is empty",status:false})
         }
         const NewLogo = await Logo.create({
-            imageUrl
+            image
         })
         res.status(201).json({message:"Add new Logo successfully",status:true,data: NewLogo})
     } catch (error) {
@@ -31,6 +32,13 @@ export const DeleteLogo = async(req,res)=>{
         const LogoId = req.params.id
         if(!LogoId) {
             return res.status(400).json({message:"LogoId is undefined",status:false})
+        }
+        const logo = await Logo.findById(LogoId)
+        if(logo && logo.active){
+            return  res.status(400).json({message:"Cannot delete active logo",status:false})
+        }
+        if(logo && logo.image && logo.image.public_id){
+            await deleteImage(logo.image.public_id)
         }
         await Logo.findByIdAndDelete(LogoId)
         res.status(200).json({message:"LogoId delete successfully",status:true})
@@ -46,8 +54,11 @@ export const ActiveLogo = async(req,res)=>{
         if(!Logoid){
             return res.status(400).json({message:"Id of Logo is empty",status:false})
         }
-        await Logo.updateMany({},{active: false},{new:true})
-        await Logo.findByIdAndUpdate(id,{active: true},{new:true})
+        await Logo.updateMany({active: true}, {active: false})
+        const activeLogo = await Logo.findByIdAndUpdate(Logoid,{active: true},{new:true})
+        if(!activeLogo){
+            return res.status(404).json({message:"Logo not existed",status:false})
+        }
         res.status(200).json({message:"Active Logo successfully",status:true})
     } catch (error) {
         console.log(error)
